# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Klinik Sehat API",
        "description": "REST API untuk aplikasi manajemen klinik. Mencakup manajemen dokter, jadwal, reservasi pasien, dan rekam medis.",
        "contact": {
            "name": "Klinik Sehat Dev Team",
            "email": "admin@klinik.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/dokters": {
            "get": {
                "tags": [
                    "Dokter"
                ],
                "summary": "Daftar semua dokter (public)",
                "description": "Menampilkan semua dokter beserta jadwal aktif. Support pencarian dan pagination.",
                "operationId": "eddd715a507d7337af6a8ecd61a72e9a",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Cari nama atau spesialisasi",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "penyakit dalam"
                        }
                    },
                    {
                        "name": "spesialisasi",
                        "in": "query",
                        "description": "Filter per spesialisasi",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Penyakit Dalam"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Nomor halaman",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Jumlah item per halaman (max 50)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Daftar dokter",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "nama": {
                                                        "type": "string",
                                                        "example": "dr. Siti Nurhaliza, Sp.PD"
                                                    },
                                                    "spesialisasi": {
                                                        "type": "string",
                                                        "example": "Penyakit Dalam"
                                                    },
                                                    "biaya_konsultasi": {
                                                        "type": "integer",
                                                        "example": 150000
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "meta": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 10
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "last_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "has_more": {
                                                    "type": "boolean",
                                                    "example": false
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Dokter"
                ],
                "summary": "Tambah dokter baru (admin only)",
                "operationId": "64533b2b394175ade9e1d39f976af2c5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nama",
                                    "spesialisasi",
                                    "no_str",
                                    "biaya_konsultasi"
                                ],
                                "properties": {
                                    "nama": {
                                        "type": "string",
                                        "example": "dr. Siti Nurhaliza, Sp.PD"
                                    },
                                    "spesialisasi": {
                                        "type": "string",
                                        "example": "Penyakit Dalam"
                                    },
                                    "no_str": {
                                        "type": "string",
                                        "example": "STR-2024-001"
                                    },
                                    "biaya_konsultasi": {
                                        "type": "integer",
                                        "example": 150000
                                    },
                                    "bio": {
                                        "type": "string",
                                        "example": "Dokter berpengalaman 10 tahun."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Dokter berhasil ditambahkan"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Bukan admin"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "bearerAuth": [],
                        "apiKeyAuth": []
                    }
                ]
            }
        },
        "/dokters/{id}": {
            "get": {
                "tags": [
                    "Dokter"
                ],
                "summary": "Detail satu dokter (public)",
                "operationId": "e51dfabf7a2595423c0702eeafadb806",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID dokter",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail dokter beserta jadwal aktif"
                    },
                    "404": {
                        "description": "Dokter tidak ditemukan"
                    }
                }
            },
            "put": {
                "tags": [
                    "Dokter"
                ],
                "summary": "Update data dokter (admin only)",
                "operationId": "fc887cfee884e609865300b8df2b72af",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "biaya_konsultasi": {
                                        "type": "integer",
                                        "example": 175000
                                    },
                                    "bio": {
                                        "type": "string",
                                        "example": "Bio diperbarui."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Dokter berhasil diperbarui"
                    },
                    "403": {
                        "description": "Bukan admin"
                    },
                    "404": {
                        "description": "Dokter tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": [],
                        "apiKeyAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Dokter"
                ],
                "summary": "Hapus dokter (admin only)",
                "operationId": "93904b002351cc0d39f19406c87c9d92",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dokter berhasil dihapus"
                    },
                    "403": {
                        "description": "Bukan admin"
                    },
                    "404": {
                        "description": "Dokter tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": [],
                        "apiKeyAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Masukkan JWT token dari response login. Format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "apiKeyAuth": {
                "type": "apiKey",
                "description": "API Key dari response login/register. Wajib untuk endpoint admin.",
                "name": "X-API-KEY",
                "in": "header"
            }
        }
    },
...
}